Replace gosu with set-priv
`gosu` is a great tool for dropping privileges, but produces false positive
CVEs in many security scanners due to being compiled with an old unsupported
version of the go compiler. gosu's stated policy is that no action
will be taken to publish releases avoiding these false positive CVEs,
and that it is left to each security scanner to update and each
user of gosu to understand the many reported CVEs.

https://github.com/tianon/gosu/blob/master/SECURITY.md#cves

Luckily, a drop in replacement `setpriv` exists in recent
versions of `util-linux`.

https://github.com/tianon/gosu?tab=readme-ov-file#setpriv

`setpriv` gives equivalent results without needing to install an
additional package and introduce unnecessary false positive go CVEs.